changed output from % to amount number

+++ src/main/java/lol/kv3rk/draft_predict/RankedSoloQ/RankedDbRequests/DTO/BestTrio.java
@@ -8,7 +8,7 @@ public interface BestTrio {
 
     String getChampion3();
 
-    double getPickRate();
+    int getPickRate();
 
     double getWinRate();
 }
